Remove the default parameter from initParamWithNodeHandle.
It's silly to have a default parameter there because we already
have an API that works without a node handle (initParam).  Thus,
remove the default nature of this.  This brings melodic back
into line with what indigo does.  I've tested all of the first
level dependencies of urdf, and none of them depend on this
behavior, so I think this will be safe to do.

Signed-off-by: Chris Lalancette <clalancette@osrfoundation.org>